Added TripCanBeChanged middleware to check if a trip is less than 24h from now and can be canceled/updated
